+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatInternalFrameUI.java
@@ -283,6 +283,9 @@ public class FlatInternalFrameUI
 
 	//---- class FlatBorderListener -------------------------------------------
 
+	/**
+	 * @since 1.6
+	 */
 	protected class FlatBorderListener
 		extends BorderListener
 	{