+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatTableUI.java
@@ -75,7 +75,7 @@ import com.formdev.flatlaf.util.UIScale;
  * @uiDefault Table.rowHeight							int
  * @uiDefault Table.showHorizontalLines					boolean
  * @uiDefault Table.showVerticalLines					boolean
- * @uiDefault Table.showLastVerticalLine				boolean
+ * @uiDefault Table.showTrailingVerticalLine			boolean
  * @uiDefault Table.intercellSpacing					Dimension
  * @uiDefault Table.selectionInactiveBackground			Color
  * @uiDefault Table.selectionInactiveForeground			Color
@@ -98,7 +98,7 @@ public class FlatTableUI
 {
 	protected boolean showHorizontalLines;
 	protected boolean showVerticalLines;
-	protected boolean showLastVerticalLine;
+	/** @since 1.6 */ protected boolean showTrailingVerticalLine;
 	protected Dimension intercellSpacing;
 
 	@Styleable protected Color selectionBackground;
@@ -135,7 +135,7 @@ public class FlatTableUI
 
 		showHorizontalLines = UIManager.getBoolean( "Table.showHorizontalLines" );
 		showVerticalLines = UIManager.getBoolean( "Table.showVerticalLines" );
-		showLastVerticalLine = UIManager.getBoolean( "Table.showLastVerticalLine" );
+		showTrailingVerticalLine = UIManager.getBoolean( "Table.showTrailingVerticalLine" );
 		intercellSpacing = UIManager.getDimension( "Table.intercellSpacing" );
 
 		selectionBackground = UIManager.getColor( "Table.selectionBackground" );
@@ -392,9 +392,10 @@ public class FlatTableUI
 	}
 
 	protected boolean hideLastVerticalLine() {
-		if( showLastVerticalLine )
+		if( showTrailingVerticalLine )
 			return false;
 
+		// do not hide if table is not a child of a scroll pane
 		Container viewport = SwingUtilities.getUnwrappedParent( table );
 		Container viewportParent = (viewport != null) ? viewport.getParent() : null;
 		if( !(viewportParent instanceof JScrollPane) )